Support missing timesteps in statistics ensemble cases (#12155)
* Support missing timesteps in statistics ensemble cases
* Adjust resolution values

+++ ApplicationLibCode/ProjectDataModel/ContourMap/RimStatisticsContourMapView.cpp
@@ -195,7 +195,7 @@ size_t RimStatisticsContourMapView::onTimeStepCountRequested()
 {
     if ( statisticsContourMap() )
     {
-        return (size_t)statisticsContourMap()->selectedTimeSteps().size();
+        return statisticsContourMap()->selectedTimeSteps().size();
     }
 
     return 0;
